refactor: committer create tree test
